requirements.txt
pyserial==3.5
adafruit-circuit-python-bno055
